Add Barbed Monsoon and Spring Command abilities

+++ include/battle_message.h
@@ -109,8 +109,8 @@
 {                                                                           \
     textVar[0] = B_BUFF_PLACEHOLDER_BEGIN;                                  \
     textVar[1] = B_BUFF_ABILITY;                                            \
-    textVar[2] = abilityId;                                                 \
-    textVar[3] = (abilityId & 0xFF00) >> 8;                                 \
+    textVar[2] = (u8)(abilityId);                                           \
+    textVar[3] = (u8)(((abilityId) & 0xFF00) >> 8);                         \
     textVar[4] = B_BUFF_EOS;                                                \
 }
